Extract GDSII byte-level reader into a submodule
Move the record framing loop, excess-64 float decode, Record/RecordData,
and the RecordData accessors out of gdsii.rs into gdsii::reader. Pure
refactor; parse() now calls the shared reader. Foundation for the tile
pipeline, which needs the same byte reader without the PcbData flattening.
