+++ crates/pcb-extract/src/parsers/gdsii.rs
@@ -4,94 +4,10 @@ use crate::error::ExtractError;
 use crate::types::*;
 use crate::ExtractOptions;
 
-// GDSII record types
-const HEADER: u8 = 0x00;
-const BGNLIB: u8 = 0x01;
-const LIBNAME: u8 = 0x02;
-const UNITS: u8 = 0x03;
-const ENDLIB: u8 = 0x04;
-const BGNSTR: u8 = 0x05;
-const STRNAME: u8 = 0x06;
-const ENDSTR: u8 = 0x07;
-const BOUNDARY: u8 = 0x08;
-const PATH: u8 = 0x09;
-const SREF: u8 = 0x0A;
-const AREF: u8 = 0x0B;
-const TEXT: u8 = 0x0C;
-const LAYER: u8 = 0x0D;
-const DATATYPE: u8 = 0x0E;
-const WIDTH: u8 = 0x0F;
-const XY: u8 = 0x10;
-const ENDEL: u8 = 0x11;
-const SNAME: u8 = 0x12;
-const COLROW: u8 = 0x13;
-const TEXTTYPE: u8 = 0x16;
-const STRING: u8 = 0x19;
-const STRANS: u8 = 0x1A;
-const MAG: u8 = 0x1B;
-const ANGLE: u8 = 0x1C;
-const PATHTYPE: u8 = 0x21;
-const BGNEXTN: u8 = 0x30;
-const ENDEXTN: u8 = 0x31;
+mod reader;
 
-/// All valid GDSII record type codes.
-const KNOWN_RECORD_TYPES: &[u8] = &[
-    0x00, // HEADER
-    0x01, // BGNLIB
-    0x02, // LIBNAME
-    0x03, // UNITS
-    0x04, // ENDLIB
-    0x05, // BGNSTR
-    0x06, // STRNAME
-    0x07, // ENDSTR
-    0x08, // BOUNDARY
-    0x09, // PATH
-    0x0A, // SREF
-    0x0B, // AREF
-    0x0C, // TEXT
-    0x0D, // LAYER
-    0x0E, // DATATYPE
-    0x0F, // WIDTH
-    0x10, // XY
-    0x11, // ENDEL
-    0x12, // SNAME
-    0x13, // COLROW
-    0x14, // TEXTNODE
-    0x15, // NODE
-    0x16, // TEXTTYPE
-    0x17, // PRESENTATION
-    0x19, // STRING
-    0x1A, // STRANS
-    0x1B, // MAG
-    0x1C, // ANGLE
-    0x1F, // REFLIBS
-    0x20, // FONTS
-    0x21, // PATHTYPE
-    0x22, // GENERATIONS
-    0x23, // ATTRTABLE
-    0x26, // ELFLAGS
-    0x2A, // NODETYPE
-    0x2B, // PROPATTR
-    0x2C, // PROPVALUE
-    0x2D, // BOX
-    0x2E, // BOXTYPE
-    0x2F, // PLEX
-    0x30, // BGNEXTN
-    0x31, // ENDEXTN
-    0x32, // TAPENUM
-    0x33, // TAPECODE
-    0x36, // FORMAT
-    0x37, // MASK
-    0x38, // ENDMASKS
-];
-
-// GDSII data types
-const DT_NONE: u8 = 0x00;
-const DT_BITARRAY: u8 = 0x01;
-const DT_I16: u8 = 0x02;
-const DT_I32: u8 = 0x03;
-const DT_F64: u8 = 0x05;
-const DT_ASCII: u8 = 0x06;
+// Byte-level record reader (record framing, float decode, RecordData accessors).
+use reader::*;
 
 /// Maximum number of flattened elements before stopping.
 const MAX_FLATTEN_ELEMENTS: usize = 500_000;
@@ -102,22 +18,6 @@ const MAX_AREF_INSTANCES: usize = 1_000_000;
 
 /// Maximum number of footprints to generate.
 const MAX_FOOTPRINTS: usize = 5_000;
-
-/// A parsed GDSII record.
-struct Record {
-    record_type: u8,
-    data: RecordData,
-}
-
-/// Data payload of a GDSII record.
-enum RecordData {
-    None,
-    BitArray(Vec<u16>),
-    Int16(Vec<i16>),
-    Int32(Vec<i32>),
-    Float64(Vec<f64>),
-    Ascii(String),
-}
 
 /// A GDSII element within a structure.
 enum GdsElement {
@@ -157,212 +57,6 @@ enum GdsElement {
 struct GdsStructure {
     name: String,
     elements: Vec<GdsElement>,
-}
-
-/// Convert GDSII excess-64 (IBM) float to IEEE 754 f64.
-///
-/// Format: 1 sign bit, 7 exponent bits (biased by 64), 56 mantissa bits.
-/// value = (-1)^sign * (mantissa / 2^56) * 16^(exponent - 64)
-fn gds_float_to_f64(bytes: &[u8]) -> f64 {
-    if bytes.len() != 8 {
-        return 0.0;
-    }
-    let sign = (bytes[0] >> 7) & 1;
-    let exponent = (bytes[0] & 0x7F) as i32;
-    let mut mantissa: u64 = 0;
-    for &b in &bytes[1..8] {
-        mantissa = (mantissa << 8) | (b as u64);
-    }
-    if mantissa == 0 {
-        return 0.0;
-    }
-    let value = (mantissa as f64 / (1u64 << 56) as f64) * 16f64.powi(exponent - 64);
-    if sign == 1 {
-        -value
-    } else {
-        value
-    }
-}
-
-/// Read a big-endian u16 from a byte slice.
-fn read_u16(data: &[u8], offset: usize) -> Result<u16, ExtractError> {
-    if offset + 2 > data.len() {
-        return Err(ExtractError::ParseError(
-            "GDSII: unexpected end of data reading u16".into(),
-        ));
-    }
-    Ok(u16::from_be_bytes([data[offset], data[offset + 1]]))
-}
-
-/// Parse all records from a GDSII byte stream.
-fn parse_records(data: &[u8]) -> Result<Vec<Record>, ExtractError> {
-    let mut records = Vec::new();
-    let mut offset = 0;
-
-    while offset < data.len() {
-        if offset + 4 > data.len() {
-            break;
-        }
-
-        let length = read_u16(data, offset)? as usize;
-        if length < 4 {
-            return Err(ExtractError::ParseError(format!(
-                "GDSII: invalid record length {} at offset {}",
-                length, offset
-            )));
-        }
-        if offset + length > data.len() {
-            return Err(ExtractError::ParseError(format!(
-                "GDSII: record at offset {} extends past end of data (length {})",
-                offset, length
-            )));
-        }
-
-        let record_type = data[offset + 2];
-        let data_type = data[offset + 3];
-
-        if !KNOWN_RECORD_TYPES.contains(&record_type) {
-            return Err(ExtractError::ParseError(format!(
-                "GDSII: unknown record type 0x{:02X} at offset {}",
-                record_type, offset
-            )));
-        }
-
-        let payload = &data[offset + 4..offset + length];
-
-        let record_data = parse_record_data(data_type, payload, record_type, offset)?;
-        records.push(Record {
-            record_type,
-            data: record_data,
-        });
-
-        offset += length;
-
-        if record_type == ENDLIB {
-            break;
-        }
-    }
-
-    Ok(records)
-}
-
-/// Parse the data payload of a record based on data type.
-fn parse_record_data(
-    data_type: u8,
-    payload: &[u8],
-    record_type: u8,
-    offset: usize,
-) -> Result<RecordData, ExtractError> {
-    match data_type {
-        DT_NONE => Ok(RecordData::None),
-        DT_BITARRAY => {
-            let mut vals = Vec::new();
-            let mut i = 0;
-            while i + 1 < payload.len() {
-                vals.push(u16::from_be_bytes([payload[i], payload[i + 1]]));
-                i += 2;
-            }
-            Ok(RecordData::BitArray(vals))
-        }
-        DT_I16 => {
-            let mut vals = Vec::new();
-            let mut i = 0;
-            while i + 1 < payload.len() {
-                vals.push(i16::from_be_bytes([payload[i], payload[i + 1]]));
-                i += 2;
-            }
-            Ok(RecordData::Int16(vals))
-        }
-        DT_I32 => {
-            let mut vals = Vec::new();
-            let mut i = 0;
-            while i + 3 < payload.len() {
-                vals.push(i32::from_be_bytes([
-                    payload[i],
-                    payload[i + 1],
-                    payload[i + 2],
-                    payload[i + 3],
-                ]));
-                i += 4;
-            }
-            Ok(RecordData::Int32(vals))
-        }
-        DT_F64 => {
-            let mut vals = Vec::new();
-            let mut i = 0;
-            while i + 7 < payload.len() {
-                vals.push(gds_float_to_f64(&payload[i..i + 8]));
-                i += 8;
-            }
-            Ok(RecordData::Float64(vals))
-        }
-        DT_ASCII => {
-            let mut s = String::from_utf8_lossy(payload).to_string();
-            // GDSII strings are padded with null bytes
-            if let Some(pos) = s.find('\0') {
-                s.truncate(pos);
-            }
-            Ok(RecordData::Ascii(s))
-        }
-        _ => Err(ExtractError::ParseError(format!(
-            "GDSII: invalid data type 0x{:02X} for record type 0x{:02X} at offset {}",
-            data_type, record_type, offset
-        ))),
-    }
-}
-
-/// Extract i16 values from record data.
-fn get_i16(data: &RecordData) -> Vec<i16> {
-    match data {
-        RecordData::Int16(v) => v.clone(),
-        _ => Vec::new(),
-    }
-}
-
-/// Extract i32 values from record data.
-fn get_i32(data: &RecordData) -> Vec<i32> {
-    match data {
-        RecordData::Int32(v) => v.clone(),
-        _ => Vec::new(),
-    }
-}
-
-/// Extract f64 values from record data.
-fn get_f64(data: &RecordData) -> Vec<f64> {
-    match data {
-        RecordData::Float64(v) => v.clone(),
-        _ => Vec::new(),
-    }
-}
-
-/// Extract ASCII string from record data.
-fn get_ascii(data: &RecordData) -> String {
-    match data {
-        RecordData::Ascii(s) => s.clone(),
-        _ => String::new(),
-    }
-}
-
-/// Extract bitarray values from record data.
-fn get_bitarray(data: &RecordData) -> Vec<u16> {
-    match data {
-        RecordData::BitArray(v) => v.clone(),
-        _ => Vec::new(),
-    }
-}
-
-/// Extract XY coordinate pairs from a record data (i32 pairs).
-fn get_xy_pairs(data: &RecordData) -> Vec<(i32, i32)> {
-    let vals = get_i32(data);
-    vals.chunks(2)
-        .filter_map(|chunk| {
-            if chunk.len() == 2 {
-                Some((chunk[0], chunk[1]))
-            } else {
-                None
-            }
-        })
-        .collect()
 }
 
 /// Parse structures from records.
